+++ CHANGELOG.md
@@ -5,6 +5,30 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2021-01-05
+
+### Added
+
+- Fully working api endpoint to get invoices by invoice number.
+- Documentation for the create and get invoice endpoints.
+
+## [1.1.1] - 2021-01-01
+
+### Added
+
+- Check if invoice doesn't already exists before creation.
+- Check if invoice type id exists. 
+
+## [1.1.0] - 2020-12-30
+
+### Added
+
+- Fully working api endpoint to store an invoince in the database.
+- Parameter validation functions for the create invoice api endpoint.
+- Unit tests for the Parameter validation functions in the databaseHelper file.
+- Database structure on project build.
+- invoice_types table automaticlly filled on build.
+
 ## [1.0.0] - 2020-12-30
 
 ### Description